refactor: Footer 분리
