+++ src/pages/mypage/styles.ts
@@ -124,40 +124,6 @@ const fundingImage = css({
   bottom: '-0.08rem',
 });
 
-const footer = css({
-  '@media (min-height: 650px)': {
-    position: 'absolute',
-    bottom: 0,
-  },
-  padding: '2.4rem 1.6rem 3.2rem',
-  width: '36rem',
-  backgroundColor: 'bg.300',
-  marginTop: '3rem',
-});
-
-const footerInfoBox = css({
-  height: '5.8rem',
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '0.6rem',
-  fontFamily: 'pretendard-regular',
-  fontSize: 'caption1',
-  color: 'text.100',
-  fontWeight: 400,
-});
-
-const termsAndConditions = css({
-  display: 'flex',
-  gap: '0.8rem',
-  fontFamily: 'pretendard-regular',
-  fontSize: 'caption1',
-  color: 'text.200',
-});
-
-const pointer = css({
-  cursor: 'pointer',
-});
-
 const cardContent = css({
   height: 'calc(100% - 4.8rem)',
   backgroundColor: 'bg.200',
@@ -251,10 +217,6 @@ export {
   goFundingSubTitle,
   goFundingTitle,
   fundingImage,
-  footer,
-  footerInfoBox,
-  termsAndConditions,
-  pointer,
   cardContent,
   flexContainer,
   iconBox,
